requirements.txt
mysql-connector-python==8.3.0
pymongo==4.6.1
python-dotenv==1.0.0